CSM removal through shadow system
